Moved Guitar component to base components dir
